@@ -4,30 +4,31 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next export",
-    "start": "node server.js"
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "@supabase/supabase-js": "^2.82.0",
-    "framer-motion": "^12.23.26",
-    "lucide-react": "^0.554.0",
-    "next": "16.0.3",
-    "razorpay": "^2.9.6",
-    "react": "19.2.0",
-    "react-dom": "19.2.0",
-    "react-hot-toast": "^2.6.0",
-    "react-icons": "^5.5.0",
-    "sonner": "^2.0.7"
+    "@supabase/supabase-js": "^2.48.0",
+    "framer-motion": "^12.0.0",
+    "lucide-react": "^0.474.0",
+    "next": "^16.1.1",
+    "razorpay": "^2.9.2",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
+    "react-hot-toast": "^2.5.1",
+    "react-icons": "^5.4.0",
+    "sonner": "^1.7.4"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
+    "@tailwindcss/postcss": "^4.0.0",
+    "@types/node": "^20.0.0",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
     "babel-plugin-react-compiler": "1.0.0",
-    "eslint": "^9",
-    "eslint-config-next": "16.0.3",
-    "tailwindcss": "^4",
-    "typescript": "^5"
+    "eslint": "^9.0.0",
+    "eslint-config-next": "15.1.6",
+    "tailwindcss": "^4.0.0",
+    "typescript": "^5.0.0"
   }
 }
